Refactor kp.Store instantiation
